+++ src/alida/cli.clj
@@ -211,8 +211,16 @@
            (map format-pruned-run pruned)))
     "Pruned 0 runs."))
 
+(defn- format-automatic-pruning
+  [{:keys [pruned_count skipped max_age_days]}]
+  (if skipped
+    "History pruning skipped because one or more indexes failed."
+    (format "History pruning removed %s runs older than %s days."
+            pruned_count
+            max_age_days)))
+
 (defn- format-crawl-result
-  [{:keys [succeeded failed]}]
+  [{:keys [succeeded failed pruning]}]
   (str/join
    \newline
    (concat
@@ -220,7 +228,8 @@
              (count succeeded)
              (count failed))]
     (map format-crawled-run succeeded)
-    (map format-failed-index failed))))
+    (map format-failed-index failed)
+    (when pruning [(format-automatic-pruning pruning)]))))
 
 (defn- crawl-exit-code
   [{:keys [failed]}]
@@ -299,7 +308,8 @@
                  #(db/prune-runs! %
                                   {:keep-last (:keep-last options)
                                    :older-than (older-than-cutoff (:older-than options))
-                                   :disabled-embeddings (:disabled-embeddings options)}))]
+                                   :disabled-embeddings (:disabled-embeddings options)
+                                   :index-names (some-> (:index options) vector)}))]
     {:exit-code 0
      :message (format-prune-result result)}))